fix(mailchimp): remove circular dependency due to scheduler, create second mailchimp scheduler
