+++ packages/backend-modules/mailchimp/lib/scheduler/index.ts
@@ -1,7 +1,5 @@
 const debug = require('debug')('mail:lib:scheduler')
 const { timeScheduler } = require('@orbiting/backend-modules-schedulers')
-import dayjs from 'dayjs'
-import { cleanedUserMailing } from './cleanedUserMailing'
 import { archiveUnsubscribed } from './archiveUnsubscribed'
 import bluebird from 'bluebird'
 
@@ -11,29 +9,6 @@ const init = async (context) => {
   debug('init')
 
   const schedulers: any[] = []
-
-  schedulers.push(
-    timeScheduler.init({
-      name: 'cleaned-user-mailing',
-      context,
-      runFunc: async (_args, context) => {
-        const { now, dryRun } = _args
-        const to = dayjs(now).format('YYYY-MM-DD')
-        const from = dayjs(now).add(-30, 'day').format('YYYY-MM-DD')
-        debug(
-          `starting job to send mail to users who were cleaned from mailing list between ${from} and ${to}`,
-        )
-        await cleanedUserMailing(from, to, context, dryRun)
-      },
-      lockTtlSecs: 60,
-      runAtTime: '09:00',
-      runAtDaysOfWeek: [1, 2, 3, 4, 5, 6, 7], // every day of the week
-      /* atm we do not send newsletters on Sundays, so there we will 
-    not have newly cleaned users but it doesn't matter to execute 
-    script anyway */
-      runInitially: DEV,
-    }),
-  )
 
   schedulers.push(
     timeScheduler.init({
